# [SPARK-18836] [CORE] Serialize one copy of task metrics in DAGScheduler