## What changes were proposed in this pull request?

Right now we serialize the empty task metrics once per task – Since this is shared across all tasks we could use the same serialized task metrics across all tasks of a stage. 

## How was this patch tested?

- [x] Run tests on EC2 to measure performance improvement