@shivaram 2016-12-13
This code was originally written by @JoshRosen. BTW while I put this up for review,  I am going to run some performance tests on EC2 today / tomm to see exactly how much benefit this gives

cc @kayousterhout