@JoshRosen 2016-12-13
Woohoo! This change has been sitting on my laptop for a while but I haven't had the time to clean it up for submission. Thanks for picking this up.