@kayousterhout 2016-12-19
Merged this to master. Thanks @shivaram!